+++ src/components/instant-apps-control-panel/test/instant-apps-control-panel.e2e.ts
@@ -1,0 +1,11 @@
+import { newE2EPage } from '@stencil/core/testing';
+
+describe('instant-apps-control-panel', () => {
+  it('renders', async () => {
+    const page = await newE2EPage();
+    await page.setContent('<instant-apps-control-panel></instant-apps-control-panel>');
+
+    const element = await page.find('instant-apps-control-panel');
+    expect(element).toHaveClass('hydrated');
+  });
+});