+++ src/components/instant-apps-filter-list/instant-apps-filter-list.tsx
@@ -21,6 +21,7 @@ const CSS = {
   numberInputContainer: 'instant-apps-filter-list__number-input-container',
   dateInputContainer: 'instant-apps-filter-list__date-picker-input-container',
   operatorDesc: 'instant-apps-filter-list__operator-description',
+  zoomTo: 'instant-apps-filter-list__zoom-to',
 };
 
 /**
@@ -105,6 +106,7 @@ export class InstantAppsFilterList {
   locale: string;
   panelEl: HTMLCalcitePanelElement;
   reactiveUtils: __esri.reactiveUtils;
+  zoomToGraphics: __esri.Graphic[];
 
   async componentWillLoad(): Promise<void> {
     this.baseClass = getMode(this.hostElement) === 'dark' ? baseClassDark : baseClassLight;
@@ -181,8 +183,9 @@ export class InstantAppsFilterList {
 
   renderFilterBlocks(layerExpression: LayerExpression): VNode {
     const filter = this.renderFilter(layerExpression);
-    const { operator } = layerExpression;
+    const { id, operator } = layerExpression;
     const operatorTranslation = operator?.trim() === 'OR' ? 'orOperator' : 'andOperator';
+    const zoomTo = this.renderZoomTo(id);
     return (
       <calcite-block
         key={layerExpression.id}
@@ -191,6 +194,7 @@ export class InstantAppsFilterList {
         open={this.openFilters ? true : undefined}
         collapsible
       >
+        {zoomTo}
         {filter}
       </calcite-block>
     );
@@ -227,10 +231,12 @@ export class InstantAppsFilterList {
   initFilterConfig(): VNode[] | undefined {
     if (this.layerExpressions?.length > 0) {
       if (this.layerExpressions.length === 1) {
-        const { operator } = this.layerExpressions[0];
+        const { id, operator } = this.layerExpressions[0];
         const operatorTranslation = operator?.trim() === 'OR' ? 'orOperator' : 'andOperator';
+        const zoomTo = this.renderZoomTo(id);
         return (
           <calcite-block class={CSS.operatorDesc} heading={this.messages?.[operatorTranslation]} open>
+            {zoomTo}
             {this.renderFilter(this.layerExpressions[0])}
           </calcite-block>
         );
@@ -319,6 +325,16 @@ export class InstantAppsFilterList {
         </calcite-button>
         <calcite-button appearance="solid" width="half" kind="brand" onClick={this.closeBtnOnClick?.bind(this)}>
           {closeText}
+        </calcite-button>
+      </div>
+    );
+  }
+
+  renderZoomTo(id: string): VNode {
+    return (
+      <div class={CSS.zoomTo}>
+        <calcite-button id={`zoom-to-${id}`} appearance="transparent" kind="neutral" scale="s" iconStart="magnifying-glass-plus" onClick={this.handleZoomTo.bind(this, id)}>
+          {this.messages?.zoomTo}
         </calcite-button>
       </div>
     );
@@ -820,6 +836,54 @@ export class InstantAppsFilterList {
         this.initDefExpressions[fl.id] = fl.definitionExpression;
       }
     });
+  }
+
+  async handleZoomTo(id: string): Promise<void> {
+    const zoomToBtn = this.panelEl.querySelector(`#zoom-to-${id}`) as HTMLCalciteButtonElement;
+    if (zoomToBtn != null) {
+      zoomToBtn.loading = true;
+    }
+    this.zoomToGraphics = [];
+    let loadingTime = 0;
+    let startGoTo = false;
+    const zoomToInterval = setInterval(() => {
+      if (loadingTime >= 1000 && startGoTo) {
+        this.view.goTo(this.zoomToGraphics);
+        if (zoomToBtn != null) {
+          zoomToBtn.loading = false;
+        }
+        clearInterval(zoomToInterval);
+      }
+      loadingTime += 500;
+    }, 500);
+    await this.getZoomToGraphics(id);
+    startGoTo = true;
+  }
+
+  async getZoomToGraphics(id: string): Promise<void> {
+    const layer = this.view.map.findLayerById(id) as FilterQueryLayer;
+    if (supportedTypes.includes(layer?.type)) {
+      const query = layer.createQuery();
+      if (layer?.capabilities?.query?.['supportsCacheHint']) {
+        query.cacheHint = true;
+      }
+      query.where = layer.definitionExpression ?? '1=1';
+      query.returnGeometry = true;
+      query.returnDistinctValues = true;
+      query.maxRecordCountFactor = 3;
+      query.returnExceededLimitFeatures = true;
+      query.outFields = [];
+      if (this.extentSelector && this.extentSelectorConfig) {
+        const geo = this.getExtent(this.extentSelector, this.extentSelectorConfig);
+        if (geo != null) query.geometry = geo;
+        query.spatialRelationship = 'intersects';
+      }
+      try {
+        const results = await layer.queryFeatures(query);
+        this.zoomToGraphics.push(...results.features);
+      } catch {}
+    }
+    return Promise.resolve();
   }
 
   generateOutput(layerId: string): void {