+++ src/components/instant-apps-header/instant-apps-header.tsx
@@ -117,6 +117,10 @@ export class InstantAppsHeader {
     reflect: true,
   })
   infoIsOpen: boolean = false;
+  /**
+   * Hover text for info button tooltip
+   */
+  @Prop() infoTitleText: string;
 
   /**
    * HTML code for custom headers.
@@ -185,9 +189,18 @@ export class InstantAppsHeader {
               {logo}
               {title}
               {this.infoButton ? (
-                <button id="infoButton" onClick={this.toggleInfo.bind(this)}>
-                  <calcite-icon icon="information-f" scale="s" />
-                </button>
+                <calcite-button
+                  style={{ '--calcite-ui-text-1': this.textColor }}
+                  id="infoButton"
+                  alignment="start"
+                  appearance="transparent"
+                  kind="neutral"
+                  icon-start="information-f"
+                  scale="s"
+                  label={this.infoTitleText}
+                  title={this.infoTitleText}
+                  onClick={this.toggleInfo.bind(this)}
+                ></calcite-button>
               ) : null}
             </span>
             <slot name="actions-end" />